requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
streamlit>=1.30.0
python-dotenv>=1.0.0